+++ server.js
@@ -1,25 +1,34 @@
 'use strict';
 
-// assign methods to variables if you're goign to use more than once
+// assign methods to variables if you're goign to use more than onnce
+// bring in all the required methods for a server
 const express = require('express');
-require('dotenv').config();
 const cors = require('cors');
 const app = express();
+
+// require 'dotenv for the environment file
+// calling the cors(), because it's required for our server
+require('dotenv').config();
 app.use(cors());
 
-const movies = require('./movies');
-const weather = require('./weather');
-
-// importing the objects from the weather.json file and assign to varible so we can use it.
-// const weatherData = require('./data/weather.json');
-// const { request } = require('express');
-
-
-
-
+// refactored movies and weather and importing from folder
+const movies = require('./functions/movies');
+const weather = require('./functions/weather');
 
 // PORT set to process.env.PORT in .env or PORT 3001 for default
 const PORT = process.env.PORT || 3002;
+
+// refactored
+app.get('/weather', weather);
+app.get('/movies', movies);
+
+app.listen(PORT, () => console.log(`Listening on ${PORT}`));
+
+
+// importing the objects from the weather.json file and assign to
+// varible so we can use it.
+// const weatherData = require('./data/weather.json');
+// const { request } = require('express');
 
 // function Movies(infor) {
 // this.title = infor.title;
@@ -40,14 +49,16 @@ const PORT = process.env.PORT || 3002;
 // lon: request.query.lon
 // })
 // .then(weatherData => {
-// response.send(weatherData.body.data.map(day => (new WeatherForecast(day))));
+// response.send(weatherData.body.data.map(day => (new WeatherForecast
+// (day))));
 // console.log(response);
 // })
 // .catch(err => (err.request, err.response));
 // });
 
 
-// https://api.themoviedb.org/3/search/movie?api_key=bc8774b1b4d020c62bb62615cad7746c&query=seattle
+// https://api.themoviedb.org/3/search/movie?
+// api_key=bc8774b1b4d020c62bb62615cad7746c&query=seattle
 // app.get('/movies', (request, response) => {
 // superagent.get('https://api.themoviedb.org/3/search/movie')
 // .query({
@@ -55,20 +66,8 @@ const PORT = process.env.PORT || 3002;
 // query: request.query.city,
 // })
 // .then(movieInfor => {
-// response.send(movieInfor.body.results.map(infor => (new Movies(infor))));
+// response.send(movieInfor.body.results.map(infor => (new Movies
+// (infor))));
 // })
 // .catch(err => (err.request, err.response));
 // });
-
-app.get('/weather', weather);
-
-app.get('/movies', movies);
-
-app.listen(PORT, () => console.log(`Listening on ${PORT}`));
-
-
-
-
-
-
-
